Separate matplotlib specific codes and plot core in Koalas
